+++ src/app/features/components/main/GoogleMapDisp.tsx
@@ -1,0 +1,10 @@
+
+const GoogleMapDisp = () => {
+    return (
+        <div className="h-[calc(100vh-20rem)] bg-muted flex items-center justify-center">
+            <p className="text-lg font-semibold">ここにGoogleマップが表示されます</p>
+        </div>
+    )
+}
+
+export default GoogleMapDisp
